+++ foundrytools_cli_2/__main__.py
@@ -1,10 +1,10 @@
 import click
 
 from foundrytools_cli_2.cli.cff import cli as cff
+from foundrytools_cli_2.cli.cmap import cli as cmap
 from foundrytools_cli_2.cli.converter import cli as converter
 from foundrytools_cli_2.cli.fix import cli as fix
 from foundrytools_cli_2.cli.font import cli as font
-from foundrytools_cli_2.cli.glyphs import cli as glyphs
 from foundrytools_cli_2.cli.gsub import cli as gsub
 from foundrytools_cli_2.cli.name import cli as name
 from foundrytools_cli_2.cli.os_2 import cli as os_2
@@ -19,10 +19,10 @@ from foundrytools_cli_2.cli.utils import cli as utils
     help="A collection of command line tools for working with font files.",
     commands={
         "cff": cff,
+        "cmap": cmap,
         "converter": converter,
         "fix": fix,
         "font": font,
-        "glyphs": glyphs,
         "gsub": gsub,
         "name": name,
         "os2": os_2,
