+++ foundrytools_cli_2/cli/cmap/__init__.py
@@ -1,0 +1,3 @@
+from foundrytools_cli_2.cli.cmap.cli import cli
+
+__all__ = ["cli"]
